Add files via upload
This program explores a maze, finding a path from an entry point to an exit one.  - The maze is stored in a text file, with # representing walls and ␣ (_empty space_) representing passages. - You’ll find examples of such mazes in the [examples](./examples) directory.     - You can also use the Maze Generator (https://github.com/ace-lectures/maze-gen) to generate others. - The Maze is surrounded by walls on its four borders, except for its entry/exit points.     - Entry and exit points are always located on the East and West border.     - The maze is not directed. As such, exit and entry can be interchanged. - At the beginning of the exploration, we're located on the entry tile, facing the opposite side (e.g., if entering by   the eastern entry, you're facing West). - The program generates a sequence of instructions to reach the opposite exit (i.e., a "path"):     - F means 'move forward' according to your current direction     - R means 'turn right' (does not move, just change direction), and L means ‘turn left’. - A canonical path contains only F, R and L symbols - -The program takes as input a maze and print the path on the standard outpu